move SamlOptions type spec
