+++ src/node-saml/types.ts
@@ -1,3 +1,5 @@
+import type { CacheProvider } from "./inmemory-cache-provider";
+
 export type SignatureAlgorithm = "sha1" | "sha256" | "sha512";
 
 export interface SamlSigningOptions {
@@ -59,4 +61,59 @@ export interface ServiceMetadataXML {
     [key: string]: XMLValue;
     SPSSODescriptor: XMLObject;
   };
+}
+
+export type RacComparision = "exact" | "minimum" | "maximum" | "better";
+
+interface SamlScopingConfig {
+  idpList?: SamlIDPListConfig[];
+  proxyCount?: number;
+  requesterId?: string[] | string;
+}
+
+/**
+ * The options required to use a SAML strategy
+ * These may be provided by means of defaults specified in the constructor
+ */
+export interface SamlOptions extends SamlSigningOptions, MandatorySamlOptions {
+  // Core
+  callbackUrl?: string;
+  path: string;
+  protocol?: string;
+  host: string;
+  entryPoint?: string;
+  issuer: string;
+  decryptionPvk?: string | Buffer;
+
+  // Additional SAML behaviors
+  additionalParams: Record<string, string>;
+  additionalAuthorizeParams: Record<string, string>;
+  identifierFormat?: string | null;
+  acceptedClockSkewMs: number;
+  attributeConsumingServiceIndex?: string;
+  disableRequestedAuthnContext: boolean;
+  authnContext: string[];
+  forceAuthn: boolean;
+  skipRequestCompression: boolean;
+  authnRequestBinding?: string;
+  racComparison: RacComparision;
+  providerName?: string;
+  passive: boolean;
+  idpIssuer?: string;
+  audience?: string;
+  scoping?: SamlScopingConfig;
+  wantAssertionsSigned?: boolean;
+
+  // InResponseTo Validation
+  validateInResponseTo: boolean;
+  requestIdExpirationPeriodMs: number;
+  cacheProvider: CacheProvider;
+
+  // Logout
+  logoutUrl: string;
+  additionalLogoutParams: Record<string, string>;
+  logoutCallbackUrl?: string;
+
+  // extras
+  disableRequestAcsUrl: boolean;
 }
