+++ test/node-saml/tests.spec.ts
@@ -6,7 +6,8 @@ import * as querystring from "querystring";
 import { parseString } from "xml2js";
 import * as fs from "fs";
 import * as sinon from "sinon";
-import { RacComparision, RequestWithUser, SamlConfig } from "../../src/passport-saml/types.js";
+import { RequestWithUser, SamlConfig } from "../../src/passport-saml/types.js";
+import { RacComparision } from "../../src/node-saml/types.js";
 import * as should from "should";
 import assert = require("assert");
 import { FAKE_CERT, TEST_CERT } from "../types";
